optimizations: material classify + indirect dispatch (Cluster 6.1)
Wires the opaque material pipelines (PBR / Unlit / Toon) to consume
the classify pass's per-shader_id tile buckets via
`dispatchWorkgroupsIndirect`. Each pipeline now runs only over tiles
its shader_id touches — the workgroup count for an empty bucket is
zero, a documented no-op. Total workgroup launches across the opaque
pass drop from `N × full_screen` to `sum_of_tiles_per_shader_id`.

* opaque main bind group: add classify output as read-only storage
  binding (slot 23). Storage count goes 8 → 9, still within the
  10/10 cap.
* opaque WGSL: `let coords = vec2<i32>(gid.xy)` replaced with a tile
  lookup — `workgroup_id.x` indexes the per-pipeline bucket;
  `local_invocation_id.xy` is the 8×8 pixel offset. Skybox handled by
  PBR pipeline only (non-PBR pipelines early-return on skybox without
  writing to avoid double-shading mixed-material tiles).
* opaque dispatch loop: always records all 3 shader_id pipelines
  (PBR always-on for skybox); dropped the empty-renderables fallback.
* classify WGSL: rename `meta` → `mesh_meta` (WGSL reserved keyword);
  inline `join32` + `U32_MAX` to avoid double-including math.wgsl.

Verified in browser with PBR / Toon / Unlit boxes — no WebGPU
validation errors, scene renders correctly across all three
specialized pipelines.

+++ crates/renderer/src/render_passes/material_classify/buffers.rs
@@ -156,6 +156,6 @@ fn write_header(dst: &mut [u8; HEADER_BYTES as usize], bucket_capacity: u32) {
 /// 1=Unlit, 2=Toon). Passed as the second arg to
 /// `dispatch_workgroups_indirect` on the material-opaque pipeline
 /// matching that shader_id.
-pub fn indirect_args_offset(bucket_index: u32) -> u64 {
-    bucket_index as u64 * INDIRECT_ARGS_STRIDE as u64
+pub fn indirect_args_offset(bucket_index: u32) -> u32 {
+    bucket_index * INDIRECT_ARGS_STRIDE
 }
